# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Company API",
        "description": "API for managing employees and companies",
        "contact": {
            "email": "krystianlata@wp.pl"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/company": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Create a new company",
                "operationId": "f4535614372edb9a76395b01983ec122",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created the company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/company/{companyId}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Show details for a specific company",
                "operationId": "3289bd3a258ed8d6e3d3824fd1a516b4",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully showed the company details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Update an existing company",
                "operationId": "916224518e6e6fbe831bc8d3734e465a",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated the company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Delete a company",
                "operationId": "60af3cc22d8d3b004ff94b112a300e11",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted the company"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            }
        },
        "/companies": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "List all companies",
                "operationId": "163ee7cd80039698933ae10ca69cdd56",
                "responses": {
                    "200": {
                        "description": "Successfully listed the companies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompanyList"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/company/{companyId}/employee": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Create a new employee for a specific company",
                "operationId": "ad22feeefb329dd884b02867820ed125",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created the employee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            }
        },
        "/employee/{employeeId}": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Show details for a specific employee",
                "operationId": "a5678358f2146bb9432b6472eaf7c164",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully showed the employee details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employee"
                ],
                "summary": "Update an employee",
                "operationId": "8f3962646e3d1ca9c6cbf9ee4265e314",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated the employee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employee"
                ],
                "summary": "Delete an employee",
                "operationId": "396ab0aa825be396e0afaf44a7277e2b",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted the employee"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            }
        },
        "/company/{companyId}/employees": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "List employees for a specific company",
                "operationId": "38476e2ad2e5d1be15281dd27c815220",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully listed the employees",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EmployeeList"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            }
        },
        "/employees": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "List all employees",
                "operationId": "17aeca6cc5a9c3d9582ddb038c1ab4b6",
                "responses": {
                    "200": {
                        "description": "Successfully listed the employees",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EmployeeList"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateCompanyRequest": {
                "required": [
                    "name",
                    "address",
                    "city",
                    "postcode",
                    "vat_number"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "postcode": {
                        "type": "string"
                    },
                    "vat_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateCompanyRequest": {
                "required": [
                    "name",
                    "vat_number",
                    "address",
                    "city",
                    "postcode"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "vat_number": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "city": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "postcode": {
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
            "CreateEmployeeRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "email"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateEmployeeRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "email"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "company_id": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Company": {
                "required": [
                    "id",
                    "name",
                    "address",
                    "city",
                    "postcode",
                    "vat_number"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "postcode": {
                        "type": "string"
                    },
                    "vat_number": {
                        "type": "string"
                    },
                    "employees": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Employee"
                        }
                    }
                },
                "type": "object"
            },
            "CompanyList": {
                "required": [
                    "id",
                    "name",
                    "vat_number"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "vat_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Employee": {
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "company": {
                        "$ref": "#/components/schemas/Company"
                    }
                },
                "type": "object"
            },
            "EmployeeList": {
                "required": [
                    "id",
                    "first_name",
                    "last_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}